backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
openai>=1.30
python-dotenv>=1.0
pypdf>=4.2
Pillow>=10.4
pytesseract>=0.3.10
youtube-transcript-api>=0.6.2
pyttsx3>=2.90
